// src/web/Scripts/Api/Client.ts
export class RegisterManagerRequest implements IRegisterManagerRequest {
    username?: string | undefined;
    password?: string | undefined;
    role?: number;
    email?: string | undefined;

    constructor(data?: IRegisterManagerRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.username = _data["username"];
            this.password = _data["password"];
            this.role = _data["role"];
            this.email = _data["email"];
        }
    }

    static fromJS(data: any): RegisterManagerRequest {
        data = typeof data === 'object' ? data : {};
        let result = new RegisterManagerRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["username"] = this.username;
        data["password"] = this.password;
        data["role"] = this.role;
        data["email"] = this.email;
        return data;
    }
}

export interface IRegisterManagerRequest {
    username?: string | undefined;
    password?: string | undefined;
    role?: number;
    email?: string | undefined;
}